Autonomous-Underwater-Vehicle---Team-UIU-H.Y.D.RA/requirements.txt
black==24.8.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
fonttools==4.53.1
importlib_metadata==8.2.0
inputs==0.5
kiwisolver==1.4.5
matplotlib==3.9.1
mypy-extensions==1.0.0
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
psutil==6.0.0
pygame==2.6.0
pyglet==2.0.17
pymongo==4.8.0
pyparsing==3.1.2
PyQt6==6.7.1
PyQt6-Charts==6.7.0
PyQt6-Charts-Qt6==6.7.2
PyQt6-Qt6==6.7.2
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.2
PyQt6-WebEngineSubwheel-Qt6==6.7.2
PyQt6_sip==13.8.0
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.16.0
tenacity==9.0.0
tomli==2.0.1
yapf==0.40.2
zipp==3.19.2
